Refactor RhinoSceneTool into partial classes
Split the 873-line RhinoSceneTool.cs into 2 files for better
maintainability and code navigation:

- RhinoSceneTool.cs: Core dispatch, ToolInfo, object operations
- RhinoSceneTool.Layers.cs: Layer CRUD actions (layers, layer_create, layer_set, layer_delete)

No functional changes - the MCP tool API remains identical.
